requirements.txt
Django
djangorestframework
beautifulsoup4
pyperclip
requests
user-agents
python-decouple